\name{SEIR-package}
\alias{SEIR-package}
\alias{SEIR}
\docType{package}
\title{
Epidemiological metapopulation markov model simulator
}
\description{
Simulate markov model using Guillespie tau leap method,
particularly SEIR epidemiological compartment model
}
\details{
\tabular{ll}{
Package: \tab SEIR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-30\cr
License: GPL-3
LazyLoad: \tab yes\cr
}
See mk.modrun and newSEIRModel for basic functionality.
}
\author{
Christian Gunning

Maintainer: Christian Gunning <yourfault@somewhere.net>
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/23782847}
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
